Move Badgeable shared examples to single spec/support/shared_examples
Solves the whole "it loaded this twice!" warning in rspec runs. And now
cookbook_spec doesn't need to directly require it. Yey!
